@@ -13,11 +13,12 @@ class LocGroup < ActiveRecord::Base
               :foreign_key => "admin_perm_id",
               :dependent => :destroy
   has_many :locations, :dependent => :destroy
-  
-  named_scope :active, :conditions => {:active => true}
-  
-  before_validation_on_create :create_permissions
-  before_validation_on_update :update_permissions
+  scope :active, :conditions => {:active => true}
+
+  before_validation(:on => :create) {:create_permissions}
+  before_validation(:on => :update) {:update_permissions}
+#  before_validation_on_create :create_permissions
+#  before_validation_on_update :update_permissions
 
   validates_presence_of :department
 
