@@ -1,21 +1,22 @@
 class Location < ActiveRecord::Base
   belongs_to :loc_group
 
-  named_scope :active, :conditions => {:active => true}
-  named_scope :in_group, 
+  scope :active, :conditions => {:active => true}
+  scope :in_group,
     lambda {|loc_group,*order| {
       :conditions => {:loc_group_id => loc_group.id},
-      :order => order.flatten.first || 'priority ASC'                                  
+      :order => order.flatten.first || 'priority ASC'
   }}
 
   has_many :time_slots
 	has_many :template_time_slots
   has_many :shifts
-	has_many :locations_requested_shifts
-	has_many :requested_shifts, :through => :locations_requested_shifts
-  has_many :locations_shift_preferences
-	has_many :shift_preferences, :through => :locations_shift_preferences
+  # has_many :locations_requested_shifts
+  # has_many :requested_shifts, :through => :locations_requested_shifts
+  #   has_many :locations_shift_preferences
+  # has_many :shift_preferences, :through => :locations_shift_preferences
   has_and_belongs_to_many :data_objects
+
 	has_and_belongs_to_many :requested_shifts
   belongs_to :category
 
@@ -74,21 +75,21 @@ class Location < ActiveRecord::Base
      ActiveRecord::Base.transaction do
         a = LocationSinksLocationSource.find(:all, :conditions => ["location_sink_type = 'Notice' AND location_source_type = 'Location' AND location_source_id = #{self.id.to_sql}"]).collect(&:location_sink_id)
         b = Link.active.collect(&:id)
-        Link.find(a & b) 
+        Link.find(a & b)
       end
   end
 
   def restrictions #TODO: this could probalby be optimized
     Restriction.current.select{|r| r.locations.include?(self)}
   end
-  
+
   def deactivate
     self.active = false
     self.save!
     #Location activation must be set prior to individual shift activation; Shift class before_save
     shifts.after_date(Time.now.utc).update_all :active => false
   end
-  
+
   def activate
     self.active = true
     self.save!
@@ -99,7 +100,7 @@ class Location < ActiveRecord::Base
         shift.active = true
       end
       shift.save
-    end    
+    end
   end
 
   def count_people_for(shift_list, min_block)
@@ -116,14 +117,14 @@ class Location < ActiveRecord::Base
     end
     people_count
   end
-  
+
   #necessary for the public cluster view
   def is_staffed_in_list?(shift_list, time)
     time = time.in_time_zone
     remaining_shifts = shift_list.select{|s| s.start <= time && (s.submitted? ? s.report.departed : s.end) >= time && !s.missed? && (s.start + s.department.department_config.grace_period.minutes <= Time.now ? (s.signed_in? || s.submitted?) : true)}
     return remaining_shifts == [] ? false : true
   end
-  
+
   #necessary for tasks
   #this really should be on the shift model I think
   def shifts_between(start_time, end_time)
@@ -131,25 +132,25 @@ class Location < ActiveRecord::Base
     end_time = end_time.to_time
     shifts = Shift.find(:all, :conditions => ["start >= #{start_time.to_sql} AND end <= #{end_time.to_sql} AND location_id = #{self.id.to_sql} AND active is true"])
   end
-  
+
   def summary_stats(start_date, end_date)
     shifts_set = shifts.on_days(start_date, end_date).active
     summary_stats = {}
-    
+
     summary_stats[:start_date] = start_date
     summary_stats[:end_date] = end_date
     summary_stats[:total] = shifts_set.size
     summary_stats[:late] = shifts_set.select{|s| s.late == true}.size
     summary_stats[:missed] = shifts_set.select{|s| s.missed == true}.size
     summary_stats[:left_early] = shifts_set.select{|s| s.left_early == true}.size
-    
+
     return summary_stats
   end
-    
+
   def detailed_stats(start_date, end_date)
     shifts_set = shifts.on_days(start_date, end_date).active
     detailed_stats = {}
-  
+
     shifts_set.each do |s|
        stat_entry = {}
        stat_entry[:id] = s.id
@@ -174,15 +175,15 @@ class Location < ActiveRecord::Base
        stat_entry[:updates_hour] = s.updates_hour
        detailed_stats[s.id] = stat_entry
     end
-    
+
     return detailed_stats
   end
-  
+
   protected
 
   def max_staff_greater_than_min_staff
     errors.add("The minimum number of staff cannot be larger than the maximum.", "") if (self.min_staff and self.max_staff and self.min_staff > self.max_staff)
   end
-  
+
 end
 
